Move classes into new files
