+++ src/srv/prayers/new_prayers_at_user.py
@@ -1,0 +1,16 @@
+from pyeo import elegant
+
+
+import datetime
+from typing import Protocol
+
+
+@elegant
+class NewPrayersAtUser(Protocol):
+    """Новые записи о статусе намаза."""
+
+    async def create(self, date: datetime.date) -> None:
+        """Создать.
+
+        :param date: datetime.date
+        """
